requirements.txt
tensorflow
keras
numpy
opencv-python-headless
pillow
pybase64
Python-IO